[#6] DB-based storage: improved auth testing structure (refactoring)
