+++ src/tests/units/auth/test_auth_dependencies.py
@@ -13,140 +13,6 @@ from src.settings import AppSettings
 from pydantic import SecretStr
 
 
-# Module-level fixtures
-@pytest.fixture
-def test_settings() -> AppSettings:
-    """Return test settings."""
-    return AppSettings(
-        api_token=SecretStr("test-token"),
-        admin_username="test-username",
-        admin_password=SecretStr("test-password"),
-        secret_key=SecretStr("test-secret-key-for-dependencies"),
-        jwt_algorithm="HS256",
-    )
-
-
-@pytest.fixture
-def mock_request() -> MagicMock:
-    """Return mock request object."""
-    request = MagicMock()
-    request.method = "GET"
-    return request
-
-
-@pytest.fixture
-def mock_make_token() -> MagicMock:
-    """Mock make_api_token function."""
-    with patch("src.modules.auth.dependencies.make_api_token") as mock:
-        mock.return_value = MagicMock(value="test-token-value", hashed_value="test-hash")
-        yield mock
-
-
-@pytest.fixture
-def mock_decode_token() -> MagicMock:
-    """Mock decode_api_token function."""
-    with patch("src.modules.auth.dependencies.decode_api_token") as mock:
-        mock.return_value = MagicMock(sub="test-user-id")
-        yield mock
-
-
-@pytest.fixture
-def mock_hash_token() -> MagicMock:
-    """Mock hash_token function."""
-    with patch("src.modules.auth.dependencies.hash_token") as mock:
-        mock.return_value = "test-hash"
-        yield mock
-
-
-@pytest.fixture
-def mock_session_uow() -> Tuple[MagicMock, AsyncMock]:
-    """Mock SASessionUOW context manager."""
-    with patch("src.modules.auth.dependencies.SASessionUOW") as mock_uow:
-        mock_session = AsyncMock()
-        mock_uow.return_value.__aenter__.return_value.session = mock_session
-        yield mock_uow, mock_session
-
-
-@pytest.fixture
-def mock_token_repository_active() -> Tuple[MagicMock, AsyncMock]:
-    """Mock TokenRepository with active token and user."""
-    with patch("src.modules.auth.dependencies.TokenRepository") as mock_repo_class:
-        mock_repo = AsyncMock()
-        mock_token = MagicMock()
-        mock_token.is_active = True
-        mock_token.user.is_active = True
-        mock_repo.get_by_token.return_value = mock_token
-        mock_repo_class.return_value = mock_repo
-        yield mock_repo_class, mock_repo
-
-
-@pytest.fixture
-def mock_token_repository_inactive_token() -> Tuple[MagicMock, AsyncMock]:
-    """Mock TokenRepository with inactive token."""
-    with patch("src.modules.auth.dependencies.TokenRepository") as mock_repo_class:
-        mock_repo = AsyncMock()
-        mock_token = MagicMock()
-        mock_token.is_active = False
-        mock_repo.get_by_token.return_value = mock_token
-        mock_repo_class.return_value = mock_repo
-        yield mock_repo_class, mock_repo
-
-
-@pytest.fixture
-def mock_token_repository_inactive_user() -> Tuple[MagicMock, AsyncMock]:
-    """Mock TokenRepository with inactive user."""
-    with patch("src.modules.auth.dependencies.TokenRepository") as mock_repo_class:
-        mock_repo = AsyncMock()
-        mock_token = MagicMock()
-        mock_token.is_active = True
-        mock_token.user.is_active = False
-        mock_repo.get_by_token.return_value = mock_token
-        mock_repo_class.return_value = mock_repo
-        yield mock_repo_class, mock_repo
-
-
-@pytest.fixture
-def mock_token_repository_unknown_token() -> Tuple[MagicMock, AsyncMock]:
-    """Mock TokenRepository with unknown token."""
-    with patch("src.modules.auth.dependencies.TokenRepository") as mock_repo_class:
-        mock_repo = AsyncMock()
-        mock_repo.get_by_token.return_value = None
-        mock_repo_class.return_value = mock_repo
-        yield mock_repo_class, mock_repo
-
-
-@pytest.fixture
-def mock_decode_token_no_identity() -> MagicMock:
-    """Mock decode_api_token function with no identity."""
-    with patch("src.modules.auth.dependencies.decode_api_token") as mock:
-        mock.return_value = MagicMock(sub="")
-        yield mock
-
-
-@pytest.fixture
-def mock_decode_token_none_identity() -> MagicMock:
-    """Mock decode_api_token function with None identity."""
-    with patch("src.modules.auth.dependencies.decode_api_token") as mock:
-        mock.return_value = MagicMock(sub=None)
-        yield mock
-
-
-@pytest.fixture
-def mock_decode_token_error() -> MagicMock:
-    """Mock decode_api_token function that raises error."""
-    with patch("src.modules.auth.dependencies.decode_api_token") as mock:
-        mock.side_effect = HTTPException(status_code=401, detail="Invalid token")
-        yield mock
-
-
-@pytest.fixture
-def mock_session_uow_error() -> MagicMock:
-    """Mock SASessionUOW context manager that raises error."""
-    with patch("src.modules.auth.dependencies.SASessionUOW") as mock_uow:
-        mock_uow.side_effect = Exception("Database error")
-        yield mock_uow
-
-
 class TestVerifyAPITokenDependency:
     """Tests for verify_api_token dependency."""
