+++ src/tests/units/auth/test_auth_edge_cases.py
